requirements.txt
sentence-transformers
pillow
vecs
numpy
argparse
psycopg2-binary
flask